LayoutTests/fast/css/parse-justify-self.html

<!DOCTYPE html>
<html>
<head>
<style>
#justifySelfBaseline {
    justify-self: baseline;
}

#justifySelfStretch {
    justify-self: stretch;
}

#justifySelfStart {
    justify-self: start;
}

#justifySelfEnd {
    justify-self: end;
}

#justifySelfCenter {
    justify-self: center;
}

#justifySelfSelfStart {
    justify-self: self-start;
}

#justifySelfSelfEnd {
    justify-self: self-end;
}

#justifySelfLeft {
    justify-self: left;
}

#justifySelfRight {
    justify-self: right;
}

#justifySelfEndUnsafe {
    justify-self: end unsafe;
}

#justifySelfCenterUnsafe {
    justify-self: center unsafe;
}

#justifySelfSelfEndSafe {
    justify-self: self-end safe;
}

#justifySelfSelfStartSafe {
    justify-self: self-start safe;
}

#justifySelfRightSafe {
    justify-self: right safe;
}

#justifySelfLeftUnsafe {
    justify-self: left unsafe;
}
</style>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div id="justifySelfBaseline"></div>
<div id="justifySelfStretch"></div>
<div id="justifySelfStart"></div>
<div id="justifySelfEnd"></div>
<div id="justifySelfCenter"></div>
<div id="justifySelfSelfStart"></div>
<div id="justifySelfSelfEnd"></div>
<div id="justifySelfLeft"></div>
<div id="justifySelfRight"></div>

<div id="justifySelfEndUnsafe"></div>
<div id="justifySelfCenterUnsafe"></div>
<div id="justifySelfSelfEndSafe"></div>
<div id="justifySelfSelfStartSafe"></div>
<div id="justifySelfRightSafe"></div>
<div id="justifySelfLeftUnsafe"></div>
<script>
description('Test that setting and getting justify-self works as expected');

debug("Test getting justify-self set through CSS");
var justifySelfBaseline = document.getElementById("justifySelfBaseline");
shouldBe("getComputedStyle(justifySelfBaseline, '').getPropertyValue('justify-self')", "'baseline'");

var justifySelfStretch = document.getElementById("justifySelfStretch");
shouldBe("getComputedStyle(justifySelfStretch, '').getPropertyValue('justify-self')", "'stretch'");

var justifySelfStart = document.getElementById("justifySelfStart");
shouldBe("getComputedStyle(justifySelfStart, '').getPropertyValue('justify-self')", "'start'");

var justifySelfEnd = document.getElementById("justifySelfEnd");
shouldBe("getComputedStyle(justifySelfEnd, '').getPropertyValue('justify-self')", "'end'");

var justifySelfCenter = document.getElementById("justifySelfCenter");
shouldBe("getComputedStyle(justifySelfCenter, '').getPropertyValue('justify-self')", "'center'");

var justifySelfSelfEnd = document.getElementById("justifySelfSelfEnd");
shouldBe("getComputedStyle(justifySelfSelfEnd, '').getPropertyValue('justify-self')", "'self-end'");

var justifySelfSelfStart = document.getElementById("justifySelfSelfStart");
shouldBe("getComputedStyle(justifySelfSelfStart, '').getPropertyValue('justify-self')", "'self-start'");

var justifySelfLeft = document.getElementById("justifySelfLeft");
shouldBe("getComputedStyle(justifySelfLeft, '').getPropertyValue('justify-self')", "'left'");

var justifySelfRight = document.getElementById("justifySelfRight");
shouldBe("getComputedStyle(justifySelfRight, '').getPropertyValue('justify-self')", "'right'");

var justifySelfEndUnsafe = document.getElementById("justifySelfEndUnsafe");
shouldBe("getComputedStyle(justifySelfEndUnsafe, '').getPropertyValue('justify-self')", "'end unsafe'");

var justifySelfCenterUnsafe = document.getElementById("justifySelfCenterUnsafe");
shouldBe("getComputedStyle(justifySelfCenterUnsafe, '').getPropertyValue('justify-self')", "'center unsafe'");

var justifySelfSelfEndSafe = document.getElementById("justifySelfSelfEndSafe");
shouldBe("getComputedStyle(justifySelfSelfEndSafe, '').getPropertyValue('justify-self')", "'self-end safe'");

var justifySelfSelfStartSafe = document.getElementById("justifySelfSelfStartSafe");
shouldBe("getComputedStyle(justifySelfSelfStartSafe, '').getPropertyValue('justify-self')", "'self-start safe'");

var justifySelfRightSafe = document.getElementById("justifySelfRightSafe");
shouldBe("getComputedStyle(justifySelfRightSafe, '').getPropertyValue('justify-self')", "'right safe'");

var justifySelfLeftUnsafe = document.getElementById("justifySelfLeftUnsafe");
shouldBe("getComputedStyle(justifySelfLeftUnsafe, '').getPropertyValue('justify-self')", "'left unsafe'");

debug("");
debug("Test initial value of justify-self through JS");
element = document.createElement("div");
document.body.appendChild(element);
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

debug("");
debug("Test getting and setting justify-self through JS");
element = document.createElement("div");
document.body.appendChild(element);
element.style.justifySelf = "center";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'center'");

element = document.createElement("div");
document.body.appendChild(element);
element.style.justifySelf = "unsafe start";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start unsafe'");

element.style.justifySelf = "start";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

debug("");
debug("Test bad combinations of justify-self");
element = document.createElement("div");
document.body.appendChild(element);
element.style.justifySelf = "start left";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "baseline safe";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "baseline center";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "stretch unsafe";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "stretch right";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "unsafe unsafe";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "unsafe";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "unsafe safe";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "center start";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "stretch unsafe";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "safe stretch";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "baseline safe";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "unsafe baseline";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "unsafe safe";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "unsafe safe left";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "unsafe left safe";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

element.style.justifySelf = "left safe unsafe safe";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

debug("");
debug("Test the value 'initial'");
element.style.justifySelf = "center";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'center'");
element.style.justifySelf = "initial";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'start'");

debug("");
debug("Test the value 'inherit'");
parentElement = document.createElement("div");
document.body.appendChild(parentElement);
parentElement.style.justifySelf = "end";
shouldBe("getComputedStyle(parentElement, '').getPropertyValue('justify-self')", "'end'");

element = document.createElement("div");
parentElement.appendChild(element);
element.style.justifySelf = "inherit";
shouldBe("getComputedStyle(element, '').getPropertyValue('justify-self')", "'end'");
</script>
</body>
</html>
